rename python stubs to decoded_array.pyi
