+++ python/zarrista/_decoded_array.pyi
